Update ALLOWED_HOSTS and add SECURE_SSL_REDIRECT
This commit updates the ALLOWED_HOSTS setting in settings.py to an empty list if DEBUG is False. It also adds the SECURE_SSL_REDIRECT setting to redirect HTTP requests to HTTPS when DEBUG is False. Additionally, some CSS changes were made to the main.css file and some HTML changes were made to the event.html file.